Refactor (create decorators, sessions files) Add catching existing user to create_user view
