+++ backend/players/views.py
@@ -1,59 +1,55 @@
-import os
 import json
 from django.contrib.auth.models import User
+from django.db import IntegrityError
 from .models import Player
 from django.views.decorators.csrf import csrf_exempt
-from functools import wraps
-from django.contrib.auth import authenticate, login, logout
-from django.core import signing
+from django.contrib.auth import authenticate
+from .decorators import session_authenticated
+from .sessions import create_encrypted_session_value
 from django.core.files.base import ContentFile
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import JsonResponse
 
 
-# Decorator factory which return a decorator for different role authorization
-def session_authenticated(required_role=None):
-    def decorator(func):
-        @wraps(func)
-        def wrapper(request, *args, **kwargs):
-            username = kwargs.get("username")
-            session_key = f"session_{username}"
-            session_value = request.COOKIES.get(session_key)
+@csrf_exempt
+def create_player(request):
+    if request.method == "POST":
+        data = json.loads(request.body)
 
-            # print("Trying to authenticate!!!")
-            # print("Session value: ", session_value)
-            if session_value:
-                session_data = decrypt_session_value(session_value)
-                print(session_data)
-                if session_data and session_data.get("is_authenticated"):
-                    # if required_role and session_data.get("role") != required_role:
-                    #     return JsonResponse({"error": "Permission denied"}, status=403)
+        # Extract required fields from the request body
+        username = data.get("username")
+        password = data.get("password")
+        display_name = data.get("displayName")
 
-                    return func(request, *args, **kwargs)
+        if not username:
+            return JsonResponse({"error": "User name is required"}, status=400)
+        if not password:
+            return JsonResponse({"error": "Password is required"}, status=400)
+        # if not display_name:
+        #     return JsonResponse({"error": "Display name is required"}, status=400)
 
+        try:
+            # Create the user
+            user = User.objects.create_user(username=username, password=password)
+            # Create the player with the associated user
+            player = Player.objects.create(user=user, display_name=display_name)
+
+            # Return success response
             return JsonResponse(
-                {"error": "Unauthorized or invalid session"}, status=401
+                {
+                    "id": player.id,
+                    "username": user.username,
+                    "displayName": player.display_name,
+                },
+                status=201,
             )
+        except IntegrityError:
+            return JsonResponse({"error": "User already exists"}, status=400)
 
-        return wrapper
+        except Exception as e:
+            return JsonResponse({"error": str(e)}, status=500)
 
-    return decorator
-
-
-# Helper function to create encrypted session value
-def create_encrypted_session_value(data):
-    signed_data = signing.dumps(data, salt=os.environ.get("AUTH_SALT"))
-    return signed_data
-
-
-# Helper function to decrypt session value
-def decrypt_session_value(signed_data):
-    try:
-        data = signing.loads(signed_data, salt=os.environ.get("AUTH_SALT"))
-        return data
-    except signing.BadSignature:
-        return None  # Invalid or tampered session data
-
+    return JsonResponse({"error": "Invalid request method"}, status=405)
 
 @csrf_exempt
 def custom_login(request):
@@ -110,45 +106,6 @@ def custom_logout(request, username):
 
 
 @csrf_exempt
-def create_player(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-
-        # Extract required fields from the request body
-        username = data.get("username")
-        password = data.get("password")
-        display_name = data.get("displayName")
-
-        if not username:
-            return JsonResponse({"error": "User name is required"}, status=400)
-        if not password:
-            return JsonResponse({"error": "Password is required"}, status=400)
-        # if not display_name:
-        #     return JsonResponse({"error": "Display name is required"}, status=400)
-
-        try:
-            # Create the user
-            user = User.objects.create_user(username=username, password=password)
-            # Create the player with the associated user
-            # player = Player.objects.create(user=user, display_name=display_name)
-            player = Player.objects.create(user=user, display_name=display_name)
-
-            # Return success response
-            return JsonResponse(
-                {
-                    "id": player.id,
-                    "username": user.username,
-                    "displayName": player.display_name,
-                },
-                status=201,
-            )
-
-        except Exception as e:
-            return JsonResponse({"error": str(e)}, status=500)
-
-    return JsonResponse({"error": "Invalid request method"}, status=405)
-
-@csrf_exempt
 @session_authenticated()
 def upload_avatar(request, username):
     if request.method == 'POST':
@@ -175,36 +132,3 @@ def upload_avatar(request, username):
         }, status=200)
 
     return JsonResponse({'error': 'Invalid request method'}, status=405)
-
-#
-# Authentication
-#
-
-
-# @csrf_exempt  # Disable CSRF just for demonstration; better to use tokens in production
-# def custom_login(request):
-#     if request.method == 'POST':
-#         # Parse the JSON data sent in the request body
-#         data = json.loads(request.body)
-
-#         username = data.get('username')
-#         password = data.get('password')
-
-#         # Authenticate the user
-#         user = authenticate(request, username=username, password=password)
-#         if user is not None:
-#             login(request, user)  # Log the user in
-#             return JsonResponse({'message': 'Login successful'}, status=200)
-#         else:
-#             return JsonResponse({'error': 'Invalid credentials'}, status=401)
-
-#     return JsonResponse({'error': 'Invalid method'}, status=400)
-
-
-# @csrf_exempt
-# def custom_logout(request):
-#     if request.method == 'POST':
-#         logout(request)  # Log the user out
-#         return JsonResponse({'message': 'Logout successful'}, status=200)
-
-#     return JsonResponse({'error': 'Invalid method'}, status=400)
